--- omf/blockmodel/__init__.py
@@ -1,10 +1,0 @@
-from .blockmodel import (
-    FreeformSubblockDefinition,
-    FreeformSubblockedModel,
-    OctreeSubblockDefinition,
-    RegularBlockModel,
-    RegularSubblockDefinition,
-    SubblockedModel,
-    TensorGridBlockModel,
-    VariableZSubblockDefinition,
-)
